requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.5.6
attrs==21.2.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
click==7.1.2
cryptography==3.4.7
distlib==0.3.1
filelock==3.0.12
flake8==3.9.1
flaky==3.7.0
identify==2.2.4
importlib-metadata==4.0.1
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.12.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pylint==2.8.2
pyparsing==2.4.7
pytest==6.2.3
python-telegram-bot==13.5
pytz==2021.1
pyupgrade==2.13.0
PyYAML==5.4.1
regex==2021.4.4
six==1.16.0
soupsieve==2.2.1
tokenize-rt==4.1.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
virtualenv==20.4.6
wrapt==1.12.1
zipp==3.4.1
